+++ R/Milestones.R
@@ -1,15 +1,23 @@
 #' Class of Milestones
 #' @description
-#' Create a class of milestone. An milestone means the time point to take an action,
-#' e.g., carry out (futility, interim, final) analysis for add/remove arms,
-#' or stop a trial early. It can also be any more general time point where trial
+#' Create a class of milestone. A milestone means the time point to take an action,
+#' e.g., carrying out (futility, interim, final) analysis for
+#' adding/removing arms, or stopping a trial early.
+#' It can also be any more general time point where trial
 #' data is used in decision making or adaptation. For example, one can define a
 #' milestone for changing randomization scheme, sample size re-assessment, trial
 #' duration extension etc.
 #'
+#' Public methods in this R6 class are used in developing
+#' this package. Thus, we have to export the whole R6 class which exposures all
+#' public methods. However, none of the public methods on this page is
+#' useful to end users. Instead, refer to the
+#' \href{https://zhangh12.github.io/TrialSimulator/articles/conditionSystem.html}{vignette}
+#' to learn how to define milestones when performing simulation using
+#' \code{TrialSimulator}.
+#'
 #' @docType class
-#' @examples
-#' ##
+#'
 #' @export
 Milestones <- R6::R6Class(
   'Milestones',
@@ -19,6 +27,7 @@ Milestones <- R6::R6Class(
     type = NULL,
     trigger_condition = NULL,
     action = NULL,
+    action_args = list(),
     triggered = FALSE, ## logical. Whether this milestone has been triggered
                       ## (to avoid repeated execution)
     silent = FALSE,
@@ -36,7 +45,12 @@ Milestones <- R6::R6Class(
     #' @param trigger_condition function to check if this milestone should
     #' trigger. See vignette \code{Condition System for Triggering Milestones in a Trial}.
     #' @param action function to execute when the milestone triggers.
-    initialize = function(name, type = name, trigger_condition, action = doNothing){
+    #' @param ... (optional) arguments of \code{action}.
+    initialize = function(name,
+                          type = name,
+                          trigger_condition,
+                          action = doNothing,
+                          ...){
       stopifnot(is.character(name) && (length(name) == 1))
       stopifnot(is.character(type))
       if(!('Condition' %in% class(trigger_condition))){
@@ -47,11 +61,41 @@ Milestones <- R6::R6Class(
       # allow no specified action, for testing purpose.
       stopifnot(is.function(action) || is.null(action))
 
+      # Capture fixed arguments for action
+      dots <- list(...)
+      if(length(dots) && (is.null(names(dots)) || any(names(dots) == ''))){
+        stop('All extra arguments to milestone(...) must be named; ',
+             'they are passed to `action`.')
+      }
+
       if(is.function(action)){
         args_in_action <- names(formals(action))
-        if(!all(c('trial', 'milestone_name') %in% args_in_action)){
-          stop('action function ', deparse(substitute(action)), ' must use ',
-               'arguments \'trial\' and \'milestone_name\'.')
+        if(length(args_in_action) == 0 || args_in_action[1] != 'trial'){
+          stop('Action function ', deparse(substitute(action)), ' must have ',
+               '\'trial\' as its first argument.')
+        }
+      }
+
+      if(!('...' %in% args_in_action)){
+        allowed_args <- setdiff(args_in_action, 'trial')
+        unknown_args <- setdiff(names(dots), allowed_args)
+        if(length(unknown_args)){
+          stop('Unknown argument(s) <',
+               paste0(unknown_args, collapse = ', '),
+               '> in the action function of milestone <',
+               name, '>. ')
+        }
+
+        req_mask <- vapply(formals(action)[allowed_args],
+                           function(x)
+                             identical(x, quote(expr=)), logical(1))
+        required <- allowed_args[req_mask]
+        missing_req <- setdiff(required, names(dots))
+        if(length(missing_req)){
+          stop('Missing required argument(s) <',
+               paste0(missing_req, collapse = ', '),
+               '> in action function of milestone <',
+               name, '>. ')
         }
       }
 
@@ -61,6 +105,7 @@ Milestones <- R6::R6Class(
       private$trigger_condition <- trigger_condition
 
       private$action <- action
+      private$action_args <- dots
       private$triggered <- FALSE
       private$is_dry_run <- FALSE
     },
@@ -103,14 +148,14 @@ Milestones <- R6::R6Class(
     execute_action = function(trial){
 
       if(private$is_dry_run){
-        action <- default_action()
+        action <- .default_action()
       }else{
-        action <- self$get_action()(trial, self$get_name())
+        action <- do.call(self$get_action(), c(list(trial), private$action_args))
       }
 
       if(!private$silent && !is.null(action)){
         message('Action for milestone <', self$get_name(), '> is executed: \n')
-        print(action)
+        # print(action)
       }
 
     },